requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
sqlmodel>=0.0.16,<0.1
python-dotenv>=1.0,<2.0
httpx>=0.27,<0.29
psycopg2-binary>=2.9,<3.0